requirements.txt
openai
tiktoken 
chromadb
langchain
gitpython